Begin making project more modular
