+++ cmd/anonymize-mysqldump/anonymize-mysqldump.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/DekodeInteraktiv/go-anonymize-mysqldump/internal/anonymize"
+)
+
+func main() {
+	anonymize.Start()
+}
